slit declare

+++ classroom/models.py
@@ -39,7 +39,7 @@ class Answer(models.Model):
 
 
 class Student(models.Model):
-    id = models.BigIntegerField(auto_created=True,primary_key=True , serialize=False, verbose_name='ID')
+    id = models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')
     user = models.OneToOneField(settings.AUTH_USER_MODEL, on_delete=models.CASCADE,default=1)
     quizzes = models.ManyToManyField(Quiz, through='TakenQuiz',blank=True)
